lots of xptcall changes: 1) OpenVMS Linux from Colin R. Blake <colin@theblakes.com>. 2) Irix from Jason Heirtzler <jasonh@m7.engr.sgi.com>. 3) Linux Alpha from Glen Nakamura <glen.nakamura@usa.net>. 4) NT Alpha from bob meader <bob@guiduck.com>. 5) x86 'this' adjust code for old or new style from Duncan Wilcox <duncan@be.com>. 6) vtbl layout test from Bert Driehuis <bert_driehuis@nl.compuware.com>. 7) improved invoke test to catch 'this' adjustment errors. 8) updated status page.

+++ xpcom/reflect/xptcall/src/md/unix/xptcinvoke_irix.cpp
@@ -65,7 +65,7 @@ invoke_copy_to_stack(PRUint64* d, PRUint32 paramCount,
            if (i < N_ARG_REGS)
               ((PRInt64*)regs)[i] = s->val.i32;
            else
-              *d++ = s->val.i64;
+              *d++ = s->val.i32;
            break;
         case nsXPTType::T_I64:
            if (i < N_ARG_REGS)
